pyproject.toml
[project]
name = "bili"
version = "0.1.0"
description = "B站UP主自动化摘要系统"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic==1.11.1",
  "faster-whisper==1.2.1",
  "feishu-sdk==1.0.3",
  "httpx[socks]==0.25.2",
  "numpy==2.4.3",
  "onnxruntime==1.24.4",
  "openai==1.3.0",
  "pillow==10.0.0",
  "pydantic==2.6.0",
  "python-dotenv==1.0.0",
  "python-telegram-bot==20.6",
  "requests==2.31.0",
  "schedule==1.2.0",
  "sqlalchemy==2.0.48",
  "yt-dlp==2025.10.14"
]
